% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{extdata}
\alias{extdata}
\title{パッケージのextdataを見つける}
\usage{
extdata(.pkg, ...)
}
\arguments{
\item{.pkg}{パッケージ名}

\item{...}{\code{list.files}のオプションに引き継がれる。フルパスが欲しいときなどに利用}
}
\description{
パッケージのextdataを見つける
}
\examples{
extdata("readr", full.names = TRUE)
}
