% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{pkg_data}
\alias{pkg_data}
\title{パッケージに同梱されているデータを読み込む}
\usage{
pkg_data(.pkg, .ptn, .f = read.csv, .dir = "/extdata", ...)
}
\arguments{
\item{.pkg}{package name}

\item{.ptn}{pattern}

\item{.f}{function which is applied to read data.}
}
\description{
依存： \code{read_data} \code{extdata}
返り値： データフレーム。
\code{.pkg} パッケージ名。
\code{.ptn} ファイル名のパターン。\code{extdata(.pkg)}でファイル名一覧を取得できる。
\code{.f = read.csv} 読み込みに利用する関数。
\code{...} 読み込み関数のオプション。
}
\examples{
pkg_data("readr", "^challenge", .f = readr::read_csv)
}
