% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{データを呼び出す}
\usage{
read_data(.dir = NULL, .ptn = NULL, .f = readxl::read_excel, ...)
}
\arguments{
\item{`.dir}{= NULL\verb{ ディレクトリ名。初期値}NULL\verb{はダウンロードフォルダを検索する。}.dir = "."\verb{を指定するとワーキングディレクトリを検索する。}.dir = "~"`を指定するとホームディレクトリを検索する。}

\item{`.ptn`}{必須。データ名のパターン}

\item{`.f}{= readxl::read_excel` 適用する関数}

\item{`...`}{\code{.f}に付与するオプション}
}
\value{
データフレームまたはデータフレームのリスト
}
\description{
パターンにマッチするファイルが2つ以上ある場合は、自動的にpurrr::mapが適用される。
}
