% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_cols.R
\name{summ_cols}
\alias{summ_cols}
\title{パネルデータのための便利な要約関数}
\usage{
summ_cols(.data, ..., .g = pref, .b = min(.data$date), .e = max(.data$date))
}
\arguments{
\item{`.data`}{データフレーム}

\item{`...`}{関数が適用される列たち。}

\item{`.g}{= pref\verb{ 行を分割するグループ。日本語が欲しい場合は}jpref`}

\item{`.b}{= min(.data$date)` 期間はじめ}

\item{`.e}{= max(.data$date)` 期間おわり}
}
\description{
任意の複数列に対して\code{mean}、\code{sd}、\code{cv}を適用する
}
