% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_fun.R
\name{summ_fun}
\alias{summ_fun}
\title{パネルデータのための便利な要約関数}
\usage{
summ_fun(.data, ..., .f, .g = pref, .b = min(.data$date), .e = max(.data$date))
}
\arguments{
\item{`.data`}{データフレーム}

\item{`...`}{関数が適用される列たち}

\item{`.f`}{列に適用する関数}

\item{`.g`}{グループ}

\item{`.b}{= min(.data$date)` 期間はじめ}

\item{`.e}{= max(.data$date)` 期間おわり}
}
\description{
任意の複数列に対して任意の関数1つを適用する
}
