% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_stats.R
\name{summ_stats}
\alias{summ_stats}
\title{パネルデータのための便利な要約関数}
\usage{
summ_stats(
  .data,
  .col,
  ...,
  .g = pref,
  .b = min(.data$date),
  .e = max(.data$date)
)
}
\arguments{
\item{`.data`}{データフレーム}

\item{`.col`}{関数が適用される列}

\item{`...`}{列に適用する関数たち}

\item{`.g`}{グループ}

\item{`.b}{= min(.data$date)` 期間はじめ}

\item{`.e}{= max(.data$date)` 期間おわり}
}
\description{
パネルデータのための便利な要約関数。
任意の1列に対して任意の複数の関数を適用する。
\code{.data} はデータフレーム。
\code{.col} は関数が適用される列。
\code{...} は列に適用する関数たち。
\code{.g} はグループ。
\code{.b = min(.data$date)} は期間はじめ。
\code{.e = max(.data$date)} は期間おわり。
}
