% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{ファイル検索のための便利な関数}
\usage{
list_files(.dir = NULL, ...)
}
\arguments{
\item{`.dir}{= NULL\verb{ ディレクトリ名。初期値}NULL\verb{はダウンロードフォルダを検索する。}.dir = "."\verb{を指定するとワーキングディレクトリを検索する。' }.dir = "~"`を指定するとホームディレクトリを検索する。}

\item{`...`}{\code{list.files}のオプションに引き継がれる}
}
\description{
ファイル検索のための便利な関数
}
