% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{データを呼び出す}
\usage{
read_data(.dir = NULL, .ptn = NULL, .f = readxl::read_excel, ...)
}
\arguments{
\item{.dir}{NULL ディレクトリ名。初期値\code{NULL}はダウンロードフォルダを検索する。\code{.dir = "."}を指定するとワーキングディレクトリを検索する。\code{.dir = "~"}を指定するとホームディレクトリを検索する。}

\item{.ptn}{必須。データ名のパターン}

\item{.f}{readxl::read_excel 適用する関数}

\item{...}{\code{.f}に付与するオプション}
}
\value{
データフレームまたはデータフレームのリスト
}
\description{
データを呼び出す。
パターンにマッチするファイルが2つ以上ある場合は、自動的に\code{purrr::map}が適用される。
返り値：データフレームまたはデータフレームのリスト
パターンにマッチするファイルが2つ以上ある場合は、自動的にpurrr::mapが適用される。
姉妹関数の\code{list_files}で目的パスのファイルを調べてから使うと便利。
\code{.dir = NULL} ディレクトリ名。初期値\code{NULL}はダウンロードフォルダを検索する。
\code{.dir = "."}を指定するとワーキングディレクトリを検索する。
\code{.dir = "~"}を指定するとホームディレクトリを検索する。
\code{.ptn} 必須。データ名のパターン
\code{.f = readxl::read_excel} 適用する関数
\code{...} \code{.f}に付与するオプション
}
\examples{
read_data(.dir = "inst/extdata", .ptn = "soukatu._2018", sheet = "Sheet1")
read_data(.ptn = "^soukatu2_2018", sheet = "Sheet1")
}
